+++ ACadSharp/Entities/DimensionAligned.cs
@@ -40,142 +40,4 @@ namespace ACadSharp.Entities
 		[DxfCodeValue(52)]
 		public double ExtLineRotation { get; set; }
 	}
-
-	/// <summary>
-	/// Represents a <see cref="DimensionAngular3Pt"/> entity.
-	/// </summary>
-	/// <remarks>
-	/// Object name <see cref="DxfFileToken.EntityDimension"/> <br/>
-	/// Dxf class name <see cref="DxfSubclassMarker.Angular3PointDimension"/>
-	/// </remarks>
-	[DxfName(DxfFileToken.EntityDimension)]
-	[DxfSubClass(DxfSubclassMarker.Angular3PointDimension)]
-	public class DimensionAngular3Pt: Dimension
-	{
-		/// <inheritdoc/>
-		public override ObjectType ObjectType => ObjectType.DIMENSION_ANG_3_Pt;
-
-		/// <inheritdoc/>
-		public override string ObjectName => DxfFileToken.EntityDimension;
-
-		/// <summary>
-		/// Definition point for linear and angular dimensions (in WCS)
-		/// </summary>
-		[DxfCodeValue(13, 23, 33)]
-		public XYZ FirstPoint { get; set; }
-
-		/// <summary>
-		/// Definition point for linear and angular dimensions (in WCS)
-		/// </summary>
-		[DxfCodeValue(14, 24, 34)]
-		public XYZ SecondPoint { get; set; }
-
-		/// <summary>
-		/// Definition point for diameter, radius, and angular dimensions (in WCS)
-		/// </summary>
-		[DxfCodeValue(15, 25, 35)]
-		public XYZ AngleVertex { get; set; }
-	}
-
-	/// <summary>
-	/// Represents a <see cref="DimensionAngular2Line"/> entity.
-	/// </summary>
-	/// <remarks>
-	/// Object name <see cref="DxfFileToken.EntityDimension"/> <br/>
-	/// Dxf class name <see cref="DxfSubclassMarker.Angular2LineDimension"/>
-	/// </remarks>
-	[DxfName(DxfFileToken.EntityDimension)]
-	[DxfSubClass(DxfSubclassMarker.Angular2LineDimension)]
-	public class DimensionAngular2Line : Dimension
-	{
-		/// <inheritdoc/>
-		public override ObjectType ObjectType => ObjectType.DIMENSION_ANG_2_Ln;
-
-		/// <inheritdoc/>
-		public override string ObjectName => DxfFileToken.EntityDimension;
-
-		/// <summary>
-		/// Definition point for linear and angular dimensions (in WCS)
-		/// </summary>
-		[DxfCodeValue(13, 23, 33)]
-		public XYZ FirstPoint { get; set; }
-
-		/// <summary>
-		/// Definition point for linear and angular dimensions (in WCS)
-		/// </summary>
-		[DxfCodeValue(14, 24, 34)]
-		public XYZ SecondPoint { get; set; }
-
-		/// <summary>
-		/// Definition point for diameter, radius, and angular dimensions (in WCS)
-		/// </summary>
-		[DxfCodeValue(15, 25, 35)]
-		public XYZ AngleVertex { get; set; }
-
-		/// <summary>
-		/// Point defining dimension arc for angular dimensions (in OCS)
-		/// </summary>
-		[DxfCodeValue(16, 26, 36)]
-		public XYZ DimensionArc { get; set; }
-	}
-
-	/// <summary>
-	/// Represents a <see cref="DimensionRadius"/> entity.
-	/// </summary>
-	/// <remarks>
-	/// Object name <see cref="DxfFileToken.EntityDimension"/> <br/>
-	/// Dxf class name <see cref="DxfSubclassMarker.RadialDimension"/>
-	/// </remarks>
-	[DxfName(DxfFileToken.EntityDimension)]
-	[DxfSubClass(DxfSubclassMarker.RadialDimension)]
-	public class DimensionRadius : Dimension
-	{
-		/// <inheritdoc/>
-		public override ObjectType ObjectType => ObjectType.DIMENSION_RADIUS;
-
-		/// <inheritdoc/>
-		public override string ObjectName => DxfFileToken.EntityDimension;
-
-		/// <summary>
-		/// Definition point for diameter, radius, and angular dimensions(in WCS)
-		/// </summary>
-		[DxfCodeValue(15, 25, 35)]
-		public XYZ AngleVertex { get; set; }
-
-		/// <summary>
-		/// Leader length for radius and diameter dimensions
-		/// </summary>
-		[DxfCodeValue(40)]
-		public double LeaderLength { get; set; }
-	}
-
-	/// <summary>
-	/// Represents a <see cref="DimensionDiameter"/> entity.
-	/// </summary>
-	/// <remarks>
-	/// Object name <see cref="DxfFileToken.EntityDimension"/> <br/>
-	/// Dxf class name <see cref="DxfSubclassMarker.DiametricDimension"/>
-	/// </remarks>
-	[DxfName(DxfFileToken.EntityDimension)]
-	[DxfSubClass(DxfSubclassMarker.DiametricDimension)]
-	public class DimensionDiameter : Dimension
-	{
-		/// <inheritdoc/>
-		public override ObjectType ObjectType => ObjectType.DIMENSION_DIAMETER;
-
-		/// <inheritdoc/>
-		public override string ObjectName => DxfFileToken.EntityDimension;
-
-		/// <summary>
-		/// Definition point for diameter, radius, and angular dimensions(in WCS)
-		/// </summary>
-		[DxfCodeValue(15, 25, 35)]
-		public XYZ AngleVertex { get; set; }
-
-		/// <summary>
-		/// Leader length for radius and diameter dimensions
-		/// </summary>
-		[DxfCodeValue(40)]
-		public double LeaderLength { get; set; }
-	}
 }
