+++ ACadSharp/Entities/DimensionAngular3Pt.cs
@@ -1,0 +1,41 @@
+﻿using ACadSharp.Attributes;
+using CSMath;
+
+namespace ACadSharp.Entities
+{
+	/// <summary>
+	/// Represents a <see cref="DimensionAngular3Pt"/> entity.
+	/// </summary>
+	/// <remarks>
+	/// Object name <see cref="DxfFileToken.EntityDimension"/> <br/>
+	/// Dxf class name <see cref="DxfSubclassMarker.Angular3PointDimension"/>
+	/// </remarks>
+	[DxfName(DxfFileToken.EntityDimension)]
+	[DxfSubClass(DxfSubclassMarker.Angular3PointDimension)]
+	public class DimensionAngular3Pt: Dimension
+	{
+		/// <inheritdoc/>
+		public override ObjectType ObjectType => ObjectType.DIMENSION_ANG_3_Pt;
+
+		/// <inheritdoc/>
+		public override string ObjectName => DxfFileToken.EntityDimension;
+
+		/// <summary>
+		/// Definition point for linear and angular dimensions (in WCS)
+		/// </summary>
+		[DxfCodeValue(13, 23, 33)]
+		public XYZ FirstPoint { get; set; }
+
+		/// <summary>
+		/// Definition point for linear and angular dimensions (in WCS)
+		/// </summary>
+		[DxfCodeValue(14, 24, 34)]
+		public XYZ SecondPoint { get; set; }
+
+		/// <summary>
+		/// Definition point for diameter, radius, and angular dimensions (in WCS)
+		/// </summary>
+		[DxfCodeValue(15, 25, 35)]
+		public XYZ AngleVertex { get; set; }
+	}
+}
